refactor: 코드 리팩토링 및 오류 수정
