+++ src/main/java/com/example/d3nserver/D3NServerApplication.java
@@ -7,9 +7,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 @SpringBootApplication
 public class D3NServerApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(D3NServerApplication.class, args);
     }
-
 }
