+++ src/main/java/com/example/d3nserver/quiz/dto/QuizDTO.java
@@ -1,10 +1,10 @@
-package com.example.d3nserver.dto;
+package com.example.d3nserver.quiz.dto;
 
-import com.example.d3nserver.news.domain.News;
 import com.example.d3nserver.quiz.domain.Quiz;
+import lombok.Data;
 
 import java.util.List;
-
+@Data
 public class QuizDTO {
 
     private String question;
@@ -12,12 +12,10 @@ public class QuizDTO {
     private Integer answer;
     private String reason;
 
-
     public QuizDTO(Quiz quiz){
         this.question = quiz.getQuestion();
         this.choiceList = quiz.getChoiceList();
         this.answer = quiz.getAnswer();
         this.reason = quiz.getReason();
-
     }
 }
