+++ src/main/java/com/example/d3nserver/quiz/repository/QuizRepository.java
@@ -8,8 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface QuizRepository extends JpaRepository<Quiz, Integer> {
-
     public List<Quiz> findAllByNews(News news);
-
-
 }
